requirements.txt
colorama==0.4.3
discord.py==1.2.5
